Prepare response variable for modeling

If the response variable is already compatible with the selected distribution, the values are preserved without modification.

Usage

prepare_response(dat, y_var, family_choice)

Arguments

dat

Data frame

y_var

Response variable

family_choice

Distribution

Value

data.frame